+++ src/ContaoEmail2UsernameBundle.php
@@ -12,9 +12,13 @@
 namespace HeimrichHannot\Email2UsernameBundle;
 
 
+use HeimrichHannot\Email2UsernameBundle\DependencyInjection\ContaoEmail2UsernameExtension;
 use Symfony\Component\HttpKernel\Bundle\Bundle;
 
-class HeimrichHannotContaoEmail2UsernameBundle extends Bundle
+class ContaoEmail2UsernameBundle extends Bundle
 {
-
+    public function getContainerExtension()
+    {
+        return new ContaoEmail2UsernameExtension();
+    }
 }