+++ src/ContaoManager/Plugin.php
@@ -17,9 +17,11 @@ use Contao\ManagerPlugin\Bundle\BundlePluginInterface;
 use Contao\ManagerPlugin\Bundle\Config\BundleConfig;
 use Contao\ManagerPlugin\Bundle\Config\ConfigInterface;
 use Contao\ManagerPlugin\Bundle\Parser\ParserInterface;
-use HeimrichHannot\Email2UsernameBundle\HeimrichHannotContaoEmail2UsernameBundle;
+use Contao\ManagerPlugin\Config\ConfigPluginInterface;
+use HeimrichHannot\Email2UsernameBundle\ContaoEmail2UsernameBundle;
+use Symfony\Component\Config\Loader\LoaderInterface;
 
-class Plugin implements BundlePluginInterface
+class Plugin implements BundlePluginInterface, ConfigPluginInterface
 {
 
     /**
@@ -32,9 +34,17 @@ class Plugin implements BundlePluginInterface
     public function getBundles(ParserInterface $parser)
     {
         return [
-            BundleConfig::create(HeimrichHannotContaoEmail2UsernameBundle::class)->setLoadAfter([
+            BundleConfig::create(ContaoEmail2UsernameBundle::class)->setLoadAfter([
                 ContaoCoreBundle::class
             ])
         ];
+    }
+
+    /**
+     * Allows a plugin to load container configuration.
+     */
+    public function registerContainerConfiguration(LoaderInterface $loader, array $managerConfig)
+    {
+        $loader->load('@ContaoEmail2UsernameBundle/Resources/config/services.yml');
     }
 }