+++ src/Resources/contao/config/config.php
@@ -3,4 +3,5 @@
 /**
  * Hooks
  */
-$GLOBALS['TL_HOOKS']['importUser'][] = ['huh.email2username.listener.hooks', 'importUserHook'];
+$GLOBALS['TL_HOOKS']['importUser']['huh_email2username'] = [
+    \HeimrichHannot\Email2UsernameBundle\EventListener\HookListener::class, 'onImportUser'];