+++ src/Resources/contao/dca/tl_user.php
@@ -5,9 +5,7 @@ $dca = &$GLOBALS['TL_DCA']['tl_user'];
 /**
  * Callbacks
  */
-$dca['config']['onsubmit_callback']['setUsernameFromEmail'] = ['huh.email2username.listener.callback', 'setUsernameFromEmail'];
-
-/**
- * Fields
- */
-System::getContainer()->get('huh.email2username.listener.callback')->modifyDca('tl_user', $dca);
+$dca['config']['onload_callback']['huh_email2username'] = [
+    \HeimrichHannot\Email2UsernameBundle\DataContainer\UserContainer::class, 'onLoad'];
+$dca['config']['onsubmit_callback']['huh_email2username'] = [
+    \HeimrichHannot\Email2UsernameBundle\DataContainer\UserContainer::class, 'onSubmit'];